Created tool to fit a fourier series (or cosine) to Hindered rotor scan data (in .txt format).
Uses Cantherm's function to perform fit, and decide which is best.
Outputs fit parameters and plot.

The data being fit need not be Potential Energy, it could be any
periodic data (such as moment of inertia of the rotor, which can also
vary with dihedral angle and must be fit to a Fourier Series for using
the software Multiwell).